Bundling soap4r & xml-mapping to fix eBay comms
We are _temporarily_ bundling the soap4r-ruby1.9 and xml-mapping gems within vendor/cache.
This has the only goal of granting us a tighter feedback loop while fixing the communication issues with eBay.
Evantually, the aim is to publish the changes and contribute them back to the community.

Here are the relevant notes regarding the gems:
* xml-mapping: I'm using the ruby1.9_compat branch of github.com/multi-io/xml-mapping ; nothing changed so far
* soap4r: I'm using the master branch of github.com/noeticpenguin/soap4r-noeticpenguin ; I had to modify lib/xsd/xmlparser/libxmlparser to properly include SaxParser